+++ amazonorders/session.py
@@ -15,7 +15,7 @@ from requests.utils import dict_from_cookiejar
 
 from amazonorders.conf import AmazonOrdersConfig, config_file_lock, cookies_file_lock, debug_output_file_lock
 from amazonorders.exception import AmazonOrdersAuthError, AmazonOrdersError, AmazonOrdersAuthRedirectError
-from amazonorders.forms import (AuthForm, CaptchaForm, JSAuthBlocker, MfaDeviceSelectForm, MfaForm,
+from amazonorders.forms import (AuthForm, AcicAuthBlocker, CaptchaForm, JSAuthBlocker, MfaDeviceSelectForm, MfaForm,
                                 SignInForm, ClaimForm, IntentForm)
 from amazonorders.util import AmazonSessionResponse, load_class
 
@@ -164,6 +164,7 @@ class AmazonSession:
                             "field-keywords"),
                 MfaForm(config,
                         config.selectors.CAPTCHA_OTP_FORM_SELECTOR),
+                AcicAuthBlocker(config),
                 JSAuthBlocker(config,
                               config.constants.JS_ROBOT_TEXT_REGEX)]